Sample console application to show how to use native functions with Semantic Kernel
